requirements.txt
torch
torchvision
torchaudio
timm
scikit-learn
numpy
pandas
Pillow
opencv-python
tqdm
pyyaml
albumentations
wandb